all renames and .mjs replace (wip 1)
