+++ api/package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "scripts": {
     "test": "jest --config=jest.config.js",
-    "start": "nodemon --experimental-modules --delay 500ms ./src/index.mjs",
-    "debug": "nodemon --experimental-modules --inspect ./src/index.mjs"
+    "start": "nodemon --experimental-modules --delay 500ms ./src/index",
+    "debug": "nodemon --experimental-modules --inspect ./src/index"
   },
   "author": "",
   "license": "ISC",
