+++ movies-app/src/components/MovieItem/MovieItem.tsx
