+++ movies-app/src/components/MovieItem/index.ts
