+++ movies-app/src/components/Spinner/index.ts
