+++ movies-app/src/store/genres/reducer.ts
